requirements.txt
mss==9.0.1
Pillow==10.0.1
PyAutoGUI==0.9.54